init.c k_init.c xa_shel.c k_main.c k_main.h xa_global.c xa_global.h: - implement shel_write(SWM_REZCHANGE) - drop last parameter for dispatch_shutdown - change event-callback-parameter from bool to short to avoid casts

+++ xaaes/src.km/xa_global.c
@@ -31,7 +31,7 @@
 struct nova_data *nova_data = NULL;
 struct cookie_mvdi mvdi_api;
 
-unsigned long next_res = 0;
+unsigned short next_res = 0;
 
 char version[32];
 char vversion[128];
